Refactor MapPage
